+++ src/ui/layout.rs
@@ -5,7 +5,7 @@ use super::graph_view::GraphView;
 #[cfg(feature = "desktop")]
 use super::keybindings::GlobalKeyHandler;
 use super::library::LibraryPanel;
-use super::paper_detail::PaperDetail;
+use super::paper_detail::{MultiSelectSummary, PaperDetail};
 use super::pdf::{PdfTabBar, PdfViewer};
 use super::sidebar::Sidebar;
 use crate::agent::types::ChatState;
@@ -94,8 +94,10 @@ pub fn Layout() -> Element {
                     _ => rsx! {
                         div { style: "flex: 1; display: flex; min-height: 0;",
                             LibraryPanel {}
-                            if lib_state.read().selected_paper_id.is_some() {
-                                PaperDetail {}
+                            match lib_state.read().selection_count() {
+                                0 => rsx! {},
+                                1 => rsx! { PaperDetail {} },
+                                _ => rsx! { MultiSelectSummary {} },
                             }
                         }
                     },